+++ core/src/terms/index.ts
@@ -1,0 +1,2 @@
+export * from "./Term";
+export * from "./listKnownTerms";